+++ src/main/java/app/controller/CreateController.java
@@ -1,0 +1,12 @@
+package app.controller;
+
+import javafx.stage.Stage;
+
+class CreateController extends Controller {
+
+  private static final String FXMLPATH = "view/createscreen.fxml";
+
+  CreateController(Stage stage) {
+    super(stage, FXMLPATH);
+  }
+}